Code improvements
To start at the start, added Bean#isWhiteListedUser(Long), this method is added for method references

Most of the Command's set methods have been renamed to add, since they don't set, but add
All the Command's methods that return a Collection of some kind have become unmodifiable + NotNull.

A collection should **never** be null, it should be empty instead

Command also received a toString, and improved the readability of isAvailableIn (imo)

The CommandHandler also received some changes. Because collections aren't null anymore, it makes the code a lot shorter and more readable.
Also fixed 2 instances where string concatenation was used instead of the logger's format capabilities.

Added some new lines between the code, replaced a bit with a Stream for readability purposes

And renamed the parameters from lambdas into what they actually are, so instead of `s`, `message` for a `Message`
Feel like this matters, since s most of the time stands for a String of some sort.